move querying and storing logic to appropriate modules
