+++ lib/active_storage.rb
@@ -1,13 +1,17 @@
-require "active_storage/relation"
-
 require "active_support/concern"
 require "active_support/inflector"
 require "active_support/core_ext/hash"
 require "active_support/core_ext/time"
 require "csv"
 
+require "active_storage/relation"
+require "active_storage/store"
+require "active_storage/querying"
+
 module ActiveStorage
   extend ActiveSupport::Concern
+
+  include Store
 
   def initialize(params = {})
     self.class.connect
@@ -17,6 +21,8 @@ module ActiveStorage
   end
 
   included do
+    extend Querying
+
     attr_reader :id
 
     private
@@ -30,62 +36,7 @@ module ActiveStorage
     end
   end
 
-  def persisted?
-    id.present?
-  end
-
-  def save
-    persisted? ? remove_old_record : generate_id
-
-    insert
-  end
-
-  def remove_old_record
-    table = CSV.table(self.class.storage_path, col_sep: ";", headers: true)
-    table.delete_if { |row| row[:id].to_s == id }
-
-    File.open(self.class.storage_path, "w") do |f|
-      f.write(table.to_csv)
-    end
-  end
-
-  private
-
-  def insert
-    CSV.open(self.class.storage_path, "ab", col_sep: ";") do |csv|
-      csv << ["id", *@@attributes].map { |attr| public_send(attr) }
-    end
-  end
-
-  def generate_id
-    self.id = Time.current.to_f
-  end
-
   class_methods do
-    def count
-      records.size
-    end
-
-    def where(attrs = {})
-      Relation.new(records).where(attrs)
-    end
-
-    def all
-      where
-    end
-
-    def records
-      connect
-
-      table = CSV.read(storage_path, col_sep: ";", headers: true)
-      records = table.entries
-
-      records.map do |record|
-        params = Hash[*record.to_a.flatten]
-        new(params)
-      end
-    end
-
     def attributes(*attrs)
       @@attributes = attrs.map(&:to_s).reject { |attr| attr == "id" }
       class_eval do
